Do not override inherited wifi attributes
Use "print config" instead of plain "print" for resources which support
it. Without this, the provider cannot distingush between attributes set
directly at the resource from inherited ones, and would convert
inherited attribures to direct overrides on apply.

Note that at the moment this works only when using Mikrotik's API
(api:// or apis://), I could not figure out how to achieve similar
functionality over the REST API.

An important consequence of this change is Terraform taking ownership
over all attributes of the affected resources. If an attribute is set
via e.g. the web UI but is not mentioned in the Terraform configuration,
then the provider will try to clear it. If that is not desired, then the
standard Terraform "lifecycle { ignore_changes = ... }" construct can be
used to tell Terraform which attributes should it leave alone.